requirements.txt
numpy==1.26.4
DrissionPage == 4.1.1.2
Pillow == 12.0.0
ddddocr == 1.5.6
requests == 2.32.5
cryptography == 46.0.3
paho-mqtt==2.1.0